ルームのサムネイル表示 (フロントエンド)
ルームのサムネイルを生成するようにいたしました。
バックエンドの変更 ( 6d37fde ) をフロントエンドに反映した形となります。
