+++ front/src/store/api/categories/index.ts
@@ -1,3 +1,3 @@
-export { CategoryAPIActionLabel, CategoryJSONType } from './constants';
 export { getLatestCategoriesAPI, getRoomCategoriesAPI } from './actions';
+export { CategoryAPIActionLabel, CategoryJSONType } from './constants';
 export { requestCategoriesAPI, requestRoomCategoriesAPI, watchCategoriesAPI } from './sagas';
