+++ front/src/store/api/rooms/index.ts
@@ -1,5 +1,6 @@
-export { RoomAPIActionLabel, RoomJSONType } from './constants';
 export { getRoomAPI, getLatestRoomsAPI, getCategoryRoomsAPI, searchRoomsAPI } from './actions';
+export { RoomAPIActionLabel, RoomJSONType } from './constants';
+export { avatarURL } from './fetch';
 export {
   requestRoomAPI,
   requestRoomsAPI,
