+++ front/src/store/index.ts
@@ -20,13 +20,11 @@ export { history } from './router';
 export { ProgressStatus } from './ui';
 export type { Channel, ProgressType } from './ui';
 
+export { useAppSelector, useAppObjectSelector, useQueryString, useRoomAvatar } from './hooks';
 export { runSaga } from './sagas';
-
-export { dispatch, dispatchPath, store, useQueryString } from './store';
+export { dispatch, dispatchPath, store } from './store';
 
 export {
-  useAppSelector,
-  useAppObjectSelector,
   logoStateSelector,
   categoryStateSelector,
   categoriesStateSelector,
